redirektovanje za placanje

+++ Bank/src/main/java/ftn/uns/ac/rs/tim6/model/PaymentRequest.java
@@ -1,7 +1,6 @@
 package ftn.uns.ac.rs.tim6.model;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.util.Date;
 
 import javax.persistence.Column;
@@ -40,7 +39,7 @@ public class PaymentRequest implements Serializable {
 	private Date merchantTimestamp;
 	
 	@Column(name = "AMOUNT")
-	private BigDecimal amount;
+	private double amount;
 	
 	@URL
 	@Column(name = "ERROR_URL")
@@ -97,11 +96,13 @@ public class PaymentRequest implements Serializable {
 		this.merchantTimestamp = merchantTimestamp;
 	}
 
-	public BigDecimal getAmount() {
+	
+
+	public double getAmount() {
 		return amount;
 	}
 
-	public void setAmount(BigDecimal amount) {
+	public void setAmount(double amount) {
 		this.amount = amount;
 	}
 
@@ -138,7 +139,7 @@ public class PaymentRequest implements Serializable {
 	}
 
 	public PaymentRequest(Long id, Integer merchantOrderId, String merchantId, String merchantPassword,
-			Date merchantTimestamp, BigDecimal amount, String errorUrl, Merchant merchant, AcquirerOrder acquirerOrder,
+			Date merchantTimestamp, long amount, String errorUrl, Merchant merchant, AcquirerOrder acquirerOrder,
 			PaymentUrlAndId paymentUrlAndId) {
 		super();
 		this.id = id;
